+++ 3rdparty/lua-5.2.4/src/llimits.h
@@ -27,7 +27,7 @@ typedef LUAI_MEM l_mem;
 typedef unsigned char lu_byte;
 
 
-#define MAX_SIZET	((size_t)(~(size_t)0)-2)
+#define MAX_SIZET	((uint32_t)(~(uint32_t)0)-2)
 
 #define MAX_LUMEM	((lu_mem)(~(lu_mem)0)-2)
