+++ 3rdparty/lua-5.2.4/src/lmem.c
@@ -23,7 +23,7 @@
 
 /*
 ** About the realloc function:
-** void * frealloc (void *ud, void *ptr, size_t osize, size_t nsize);
+** void * frealloc (void *ud, void *ptr, uint32_t osize, uint32_t nsize);
 ** (`osize' is the old size, `nsize' is the new size)
 **
 ** * frealloc(ud, NULL, x, s) creates a new block of size `s' (no
@@ -43,7 +43,7 @@
 #define MINSIZEARRAY	4
 
 
-void *luaM_growaux_ (lua_State *L, void *block, int *size, size_t size_elems,
+void *luaM_growaux_ (lua_State *L, void *block, int *size, uint32_t size_elems,
                      int limit, const char *what) {
   void *newblock;
   int newsize;
@@ -72,10 +72,10 @@ l_noret luaM_toobig (lua_State *L) {
 /*
 ** generic allocation routine.
 */
-void *luaM_realloc_ (lua_State *L, void *block, size_t osize, size_t nsize) {
+void *luaM_realloc_ (lua_State *L, void *block, uint32_t osize, uint32_t nsize) {
   void *newblock;
   global_State *g = G(L);
-  size_t realosize = (block) ? osize : 0;
+  uint32_t realosize = (block) ? osize : 0;
   lua_assert((realosize == 0) == (block == NULL));
 #if defined(HARDMEMTESTS)
   if (nsize > realosize && g->gcrunning)
