+++ 3rdparty/lua-5.2.4/src/lstring.c
@@ -31,7 +31,7 @@
 ** equality for long strings
 */
 int luaS_eqlngstr (TString *a, TString *b) {
-  size_t len = a->tsv.len;
+  uint32_t len = a->tsv.len;
   lua_assert(a->tsv.tt == LUA_TLNGSTR && b->tsv.tt == LUA_TLNGSTR);
   return (a == b) ||  /* same instance or... */
     ((len == b->tsv.len) &&  /* equal length and ... */
@@ -48,10 +48,10 @@ int luaS_eqstr (TString *a, TString *b) {
 }
 
 
-unsigned int luaS_hash (const char *str, size_t l, unsigned int seed) {
+unsigned int luaS_hash (const char *str, uint32_t l, unsigned int seed) {
   unsigned int h = seed ^ cast(unsigned int, l);
-  size_t l1;
-  size_t step = (l >> LUAI_HASHLIMIT) + 1;
+  uint32_t l1;
+  uint32_t step = (l >> LUAI_HASHLIMIT) + 1;
   for (l1 = l; l1 >= step; l1 -= step)
     h = h ^ ((h<<5) + (h>>2) + cast_byte(str[l1 - 1]));
   return h;
@@ -95,10 +95,10 @@ void luaS_resize (lua_State *L, int newsize) {
 /*
 ** creates a new string object
 */
-static TString *createstrobj (lua_State *L, const char *str, size_t l,
+static TString *createstrobj (lua_State *L, const char *str, uint32_t l,
                               int tag, unsigned int h, GCObject **list) {
   TString *ts;
-  size_t totalsize;  /* total size of TString object */
+  uint32_t totalsize;  /* total size of TString object */
   totalsize = sizeof(TString) + ((l + 1) * sizeof(char));
   ts = &luaC_newobj(L, tag, totalsize, list, 0)->ts;
   ts->tsv.len = l;
@@ -113,7 +113,7 @@ static TString *createstrobj (lua_State *L, const char *str, size_t l,
 /*
 ** creates a new short string, inserting it into string table
 */
-static TString *newshrstr (lua_State *L, const char *str, size_t l,
+static TString *newshrstr (lua_State *L, const char *str, uint32_t l,
                                        unsigned int h) {
   GCObject **list;  /* (pointer to) list where it will be inserted */
   stringtable *tb = &G(L)->strt;
@@ -130,7 +130,7 @@ static TString *newshrstr (lua_State *L, const char *str, size_t l,
 /*
 ** checks whether short string exists and reuses it or creates a new one
 */
-static TString *internshrstr (lua_State *L, const char *str, size_t l) {
+static TString *internshrstr (lua_State *L, const char *str, uint32_t l) {
   GCObject *o;
   global_State *g = G(L);
   unsigned int h = luaS_hash(str, l, g->seed);
@@ -153,7 +153,7 @@ static TString *internshrstr (lua_State *L, const char *str, size_t l) {
 /*
 ** new string (with explicit length)
 */
-TString *luaS_newlstr (lua_State *L, const char *str, size_t l) {
+TString *luaS_newlstr (lua_State *L, const char *str, uint32_t l) {
   if (l <= LUAI_MAXSHORTLEN)  /* short string? */
     return internshrstr(L, str, l);
   else {
@@ -172,7 +172,7 @@ TString *luaS_new (lua_State *L, const char *str) {
 }
 
 
-Udata *luaS_newudata (lua_State *L, size_t s, Table *e) {
+Udata *luaS_newudata (lua_State *L, uint32_t s, Table *e) {
   Udata *u;
   if (s > MAX_SIZET - sizeof(Udata))
     luaM_toobig(L);
