+++ 3rdparty/lua-5.2.4/src/lstring.h
@@ -34,12 +34,12 @@
 #define eqshrstr(a,b)	check_exp((a)->tsv.tt == LUA_TSHRSTR, (a) == (b))
 
 
-LUAI_FUNC unsigned int luaS_hash (const char *str, size_t l, unsigned int seed);
+LUAI_FUNC unsigned int luaS_hash (const char *str, uint32_t l, unsigned int seed);
 LUAI_FUNC int luaS_eqlngstr (TString *a, TString *b);
 LUAI_FUNC int luaS_eqstr (TString *a, TString *b);
 LUAI_FUNC void luaS_resize (lua_State *L, int newsize);
-LUAI_FUNC Udata *luaS_newudata (lua_State *L, size_t s, Table *e);
-LUAI_FUNC TString *luaS_newlstr (lua_State *L, const char *str, size_t l);
+LUAI_FUNC Udata *luaS_newudata (lua_State *L, uint32_t s, Table *e);
+LUAI_FUNC TString *luaS_newlstr (lua_State *L, const char *str, uint32_t l);
 LUAI_FUNC TString *luaS_new (lua_State *L, const char *str);
